TP2 and  3/services/epf-flower-data-science/requirements.txt
gunicorn~=20.1
uvicorn==0.17.6
fastapi==0.95.1
fastapi-utils==0.2.1
pydantic==1.10
opendatasets
pytest
google-cloud-firestore